+++ fuel-vm/src/interpreter/executors/instruction/tests/reserved_registers.rs
@@ -75,7 +75,7 @@ fn cant_write_to_reserved_registers(raw_random_instruction: u32) -> TestResult {
     let tx = tx
         .into_checked(block_height, &consensus_params)
         .expect("failed to check tx")
-        .into_ready(zero_gas_price, vm.gas_costs(), &fee_params)
+        .into_ready(zero_gas_price, vm.gas_costs(), &fee_params, None)
         .expect("failed dynamic checks");
 
     vm.init_script(tx).expect("Failed to init VM");