+++ fuel-vm/src/transactor.rs
@@ -214,7 +214,7 @@ where
         let fee_params = self.interpreter.fee_params();
 
         let ready = checked
-            .into_ready(gas_price, gas_costs, fee_params)
+            .into_ready(gas_price, gas_costs, fee_params, None)
             .map_err(InterpreterError::CheckError)?;
 
         self.deploy_ready_tx(ready)
@@ -238,7 +238,7 @@ where
         let fee_params = self.interpreter.fee_params();
 
         let ready = checked
-            .into_ready(gas_price, gas_costs, fee_params)
+            .into_ready(gas_price, gas_costs, fee_params, None)
             .map_err(InterpreterError::CheckError)?;
 
         self.execute_ready_upgrade_tx(ready)
@@ -262,7 +262,7 @@ where
         let fee_params = self.interpreter.fee_params();
 
         let ready = checked
-            .into_ready(gas_price, gas_costs, fee_params)
+            .into_ready(gas_price, gas_costs, fee_params, None)
             .map_err(InterpreterError::CheckError)?;
 
         self.execute_ready_upload_tx(ready)
@@ -286,7 +286,7 @@ where
         let fee_params = self.interpreter.fee_params();
 
         let ready = checked
-            .into_ready(gas_price, gas_costs, fee_params)
+            .into_ready(gas_price, gas_costs, fee_params, None)
             .map_err(InterpreterError::CheckError)?;
 
         self.execute_ready_blob_tx(ready)
@@ -314,9 +314,10 @@ where
         let gas_price = self.interpreter.gas_price();
         let gas_costs = self.interpreter.gas_costs();
         let fee_params = self.interpreter.fee_params();
+        let block_height = self.interpreter.context().block_height();
 
         match tx
-            .into_ready(gas_price, gas_costs, fee_params)
+            .into_ready(gas_price, gas_costs, fee_params, block_height)
             .map_err(InterpreterError::CheckError)
         {
             Ok(ready_tx) => self.transact_ready_tx(ready_tx),